requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
requests>=2.31.0
beautifulsoup4>=4.12.0
email-validator==2.1.0.post1
textblob==0.17.1
tldextract==5.1.1 